Add Quirks for UA on iPad for tiktok.com (macOS version) https://bugs.webkit.org/show_bug.cgi?id=291113 <rdar://148467973>
Reviewed by Wenson Hsieh.

This patch builds on the work in Bug 290372 to ensure the same behavior on macOS
(as well as other WebKit ports).

It also adjust the flow in the iOS case to allow the quirk to be turned off if the
user deactivated custom website quirks.

* Source/WebKit/UIProcess/API/APIWebsitePolicies.h: Add a Move operator for setCustomUserAgent
* Source/WebCore/page/Quirks.cpp:
(WebCore::Quirks::needsCustomUserAgentOverride): Do more of the work in this method to
avoid duplicating it for macOS..
* Source/WebCore/page/Quirks.h:
* Source/WebKit/UIProcess/WebPageProxy.cpp: Adjust compile macro now that both iOS and
macOS have an implementation of this.
* Source/WebKit/UIProcess/ios/WebPageProxyIOS.mm:
(WebKit::WebPageProxy::effectiveContentModeAfterAdjustingPolicies): Honor the 'Enable
Website Quirks' setting and adopt the Move overload.
* Source/WebKit/UIProcess/mac/WebPageProxyMac.mm:
(WebKit::WebPageProxy::effectiveContentModeAfterAdjustingPolicies): Add logic for the
Quirk on macOS and adopt the Move overload.

Canonical link: https://commits.webkit.org/293441@main