`DataListButtonElement` should override `isDisabledFormControl()` to report the shadow host's state https://bugs.webkit.org/show_bug.cgi?id=291311 rdar://148877381
Reviewed by Ryosuke Niwa.

Override `isDisabledFormControl()` for DataListButtonElements
to report the disabled state of the shadow host, if available.
Updated implementation of `isDisabledFormControl()` for
SpinButtonElements to match (behavior is the same).

Added LayoutTest to ensure that, when a disabled list button is
pressed, it appears the same as an unpressed disabled list button.

* LayoutTests/fast/forms/datalist/datalist-button-disabled-pressed-expected.html: Added.
* LayoutTests/fast/forms/datalist/datalist-button-disabled-pressed.html: Added.
* Source/WebCore/html/shadow/DataListButtonElement.cpp:
(WebCore::DataListButtonElement::isDisabledFormControl const):
* Source/WebCore/html/shadow/DataListButtonElement.h:
* Source/WebCore/html/shadow/SpinButtonElement.cpp:
(WebCore::SpinButtonElement::isDisabledFormControl const):
* Source/WebCore/html/shadow/SpinButtonElement.h:

Canonical link: https://commits.webkit.org/293504@main