Attribute prefetched DNS requests to host app https://bugs.webkit.org/show_bug.cgi?id=291333 rdar://146158222
Reviewed by Alex Christensen.

In 290327@main we enabled `<link rel="dns-prefetch">` on iOS. This caused regressions on some power
benchmarks that measure how much networking is attributed to each process, because we don't
currently attribute prefetched DNS requests to the host app. This adds the appropriate attribution
so that those requests are attributed to the app rather than to NetworkProcess.

* Source/WebCore/platform/network/cf/DNSResolveQueueCFNet.cpp:
(WebCore::DNSResolveQueueCFNet::performDNSLookup):

Canonical link: https://commits.webkit.org/293525@main