REGRESSION(292734@main): Fix build break on macOS Ventura https://bugs.webkit.org/show_bug.cgi?id=291378 rdar://148993362
Reviewed by Abrar Rahman Protyasha.

In 286164@main, the calls to lambda isHDRColorSpace() was replaced by calling a
new function called DestinationColorSpace::usesStandardRange(). This replacement
for cleaning up the code and delete repeated calls to CG APIs.

But this new function was deleted in 292734@main. This caused a compilation error
on macOS Ventura. So restore the check for HDR in NativeImage::draw() as it was
before 286164@main.

* Source/WebCore/platform/graphics/cg/NativeImageCG.cpp:
(WebCore::NativeImage::draw):

Canonical link: https://commits.webkit.org/293531@main