GPUP: Frequent RemoteImageBuffer::PutPixelBuffer blocks on connection semaphore signal https://bugs.webkit.org/show_bug.cgi?id=289982 rdar://problem/147331680
Reviewed by Simon Fraser.

Fix by marking the PutPixelBuffer as batched. This reduces the semaphore
signals. If the pixel buffer is large, start the command rightaway.

Moves the PutPixelBuffer to RemoteImageBufferProxy, as that is the
place that sends the RemoteImageBuffer messages.

* Source/WebKit/GPUProcess/graphics/RemoteImageBuffer.messages.in:
* Source/WebKit/Platform/IPC/StreamClientConnection.h:
* Source/WebKit/WebProcess/GPU/graphics/RemoteImageBufferProxy.cpp:
(WebKit::RemoteImageBufferProxy::putPixelBuffer):
* Source/WebKit/WebProcess/GPU/graphics/RemoteRenderingBackendProxy.cpp:
(WebKit::RemoteRenderingBackendProxy::putPixelBufferForImageBuffer): Deleted.
* Source/WebKit/WebProcess/GPU/graphics/RemoteRenderingBackendProxy.h:

Canonical link: https://commits.webkit.org/293533@main