[Page color sampling] Fixed-position color extension views disappear when rubber-banding https://bugs.webkit.org/show_bug.cgi?id=291399 rdar://148840634
Reviewed by Abrar Rahman Protyasha.

Teach `fixedContainerEdges` to handle rubber-banding on macOS, without failing to hit-test to fixed-
position elements on the rubber-banded edge. Currently, the `hitTest` fails and returns `false`, due
to the fact that the `rectForFixedPositionLayout` gets pushed outside of the document rect (e.g.
when scrolling against the top, the origin has a negative `y`). Simply intersecting this rect for
fixed-position layout with the document rect is insufficient, as the hit-test will still fail and
hit the body if the rubber-banding distance is large enough that the fixed container's layer bounds
lie outside of the document rect.

As a workaround, when rubber-banding, we can instead hit-test using a rect that extends from the
midpoint of the rect edge on the rect for fixed-position layout, to the midpoint of the document
rect (with a slight inset margin). This allows hit-testing to properly find fixed-position elements
whose clip rects lie outside of the document bounds when rubber-banding.

* LayoutTests/fast/page-color-sampling/color-sampling-while-rubber-banding-expected.txt: Added.
* LayoutTests/fast/page-color-sampling/color-sampling-while-rubber-banding.html: Added.
* Source/WebCore/page/LocalFrameView.cpp:
(WebCore::LocalFrameView::fixedContainerEdges const):

Canonical link: https://commits.webkit.org/293575@main