Skad network 4.0 web ad get-signed-payload request stopped working https://bugs.webkit.org/show_bug.cgi?id=290833 rdar://148430376
Reviewed by John Wilander.

We already had infrastructure to transfer a PrivateClickMeasurement from an opener
to an opened page, but when the navigation happens in a new process without an opener
relationship, such as when target=_blank is used on an anchor tag, we weren't transferring
it.  This restores the transfer and verifies its success with a unit test.

* Source/WebKit/UIProcess/WebPageProxy.cpp:
(WebKit::WebPageProxy::createNewPage):
* Tools/TestWebKitAPI/Tests/WebKitCocoa/EventAttribution.mm:
(TestWebKitAPI::setupSKAdNetworkTest):
(TestWebKitAPI::TEST(PrivateClickMeasurement, SKAdNetwork)):
(TestWebKitAPI::TEST(PrivateClickMeasurement, SKAdNetworkAboutBlank)):
(TestWebKitAPI::TEST(PrivateClickMeasurement, SKAdNetworkWithoutNavigatingToAppStoreLink)):

Canonical link: https://commits.webkit.org/293594@main