[Page color sampling] netflix.com: top navigation header sometimes doesn't get color extension after scrolling https://bugs.webkit.org/show_bug.cgi?id=291447 rdar://149028791
Reviewed by Aditya Keerthi.

Ignore `pointer-events: none;` when hit-testing for fixed-position containers near the edges of the
viewport, for color sampling. Even if a fixed/stickly-position container is hidden for the purposes
of hit-testing for user interaction, we still want to be able to extend its background colors if
needed, if it's still visible to the user.

* LayoutTests/fast/page-color-sampling/basic-fixed-container-edge-sampling.html:

Augment this basic layout test to cover this `pointer-events: none;` scenario as well.

* Source/WebCore/page/LocalFrameView.cpp:
(WebCore::LocalFrameView::fixedContainerEdges const):

Canonical link: https://commits.webkit.org/293604@main