[JSC] Handle unknown annotations in Temporal strings https://bugs.webkit.org/show_bug.cgi?id=223166
Reviewed by Yusuke Suzuki.

This is part of a normative change to the Temporal proposal:
https://github.com/tc39/proposal-temporal/pull/2397
This change stops assuming that all calendar annotations have the `u-ca`
key, and instead parses the key according to the grammar in the spec
text. Unknown annotations are ignored unless they have the `!` flag, in
which case they cause the string to throw.

* JSTests/stress/temporal-instant.js:
* Source/JavaScriptCore/runtime/ISO8601.cpp:
(JSC::ISO8601::canBeCalendar):
(JSC::ISO8601::parseCalendar):
* Source/JavaScriptCore/runtime/ISO8601.h:

Canonical link: https://commits.webkit.org/293705@main