SWServer can create its background fetch engine within its destructor rdar://145610350 https://bugs.webkit.org/show_bug.cgi?id=291516
Reviewed by Chris Dumez.

SWServer can be destroyed as part of NetworkSession destructor.
But the background fetch engine can be created at that time when clearing a registration.
To prevent this, we are now closing the SWServer when the NetworkSession is destroyed.

* Source/WebCore/workers/service/server/SWServer.cpp:
(WebCore::SWServer::~SWServer):
(WebCore::SWServer::close):
* Source/WebCore/workers/service/server/SWServer.h:
* Source/WebKit/NetworkProcess/NetworkSession.cpp:
(WebKit::NetworkSession::invalidateAndCancel):

Canonical link: https://commits.webkit.org/293721@main