Optimize AffineTransform::mapRect() https://bugs.webkit.org/show_bug.cgi?id=291635 rdar://149393231
Reviewed by Kimmo Kinnunen.

The existing code maps each corner of the rect through the transform, forms a quad
from the result, and computes the bounds of the quad.

We can optimize this since the x and y values of the corner points are shared, so
we can save some computation.

* Source/WebCore/platform/graphics/transforms/AffineTransform.cpp:
(WebCore::AffineTransform::mapRect const):

Canonical link: https://commits.webkit.org/293798@main