[Screen Time Refactoring] Release logging is misleading https://bugs.webkit.org/show_bug.cgi?id=291704 rdar://149502357
Reviewed by Aditya Keerthi.

Currently, the release log to aid in debugging Screen Time
is causing confusing and is misleading.

For example, "Screen Time has updated visibility to show system shield"
does not mean that a Screen Time shield is up right now. It is
supposed to convey that if there is a blocked URL, we are
configured to show the system shield in the blocked case.

Thus, change release logging to be avoid confusion.

* Source/WebKit/UIProcess/API/Cocoa/WKWebView.mm:
(-[WKWebView _updateScreenTimeBasedOnWindowVisibility]):

Canonical link: https://commits.webkit.org/293834@main