ServiceWorkerRegistryData is often copied to completion handlers when caller is attempting to Move the data https://bugs.webkit.org/show_bug.cgi?id=291701 rdar://149497889
Reviewed by Sihui Liu.

The signature for `importServiceWorkerRegistrations` takes a CompletionHandler that takes
`std::optional<Vector<WebCore::ServiceWorkerContextData>>` as its argument. This requires a copy of
the Vector of items.

The only caller of this API attempts to Move this vector, so the signature should allow the move.

* Source/WebCore/workers/service/server/SWRegistrationStore.h: Indicate move operation.
* Source/WebKit/NetworkProcess/ServiceWorker/WebSWRegistrationStore.cpp:
(WebKit::WebSWRegistrationStore::importRegistrations): Ditto.
* Source/WebKit/NetworkProcess/ServiceWorker/WebSWRegistrationStore.h:
* Source/WebKit/NetworkProcess/storage/NetworkStorageManager.cpp:
(WebKit::NetworkStorageManager::importServiceWorkerRegistrations): Ditto.
* Source/WebKit/NetworkProcess/storage/NetworkStorageManager.h:

Canonical link: https://commits.webkit.org/293859@main