[JSC] Improve ISO8601::CalendarRecord struct https://bugs.webkit.org/show_bug.cgi?id=223166
Reviewed by Justin Michaud and Sosuke Suzuki.

Follow-up to comment about using an enum for the m_unknown member of
CalendarRecord. See
https://github.com/WebKit/WebKit/pull/44103#discussion_r2045376760

The annotation syntax is defined in RFC 9557, so let's call the struct
RFC9557Annotation. We define type aliases for the flag, the key, and the
value.

Once the annotations are parsed and the unknown ones are discarded, we
no longer need to pass around the whole record with flag and key
members. We can just pass the RFC9557Value, which we then type-alias to
CalendarID for clarity.

* Source/JavaScriptCore/runtime/ISO8601.cpp:
(JSC::ISO8601::parseOneCalendar):
(JSC::ISO8601::parseCalendar):
(JSC::ISO8601::parseCalendarTime):
(JSC::ISO8601::parseCalendarDateTime):
* Source/JavaScriptCore/runtime/ISO8601.h:

Canonical link: https://commits.webkit.org/293864@main