Fix iOS touch events with site isolation on after 290019@main rdar://148463266 https://bugs.webkit.org/show_bug.cgi?id=291752
Reviewed by Brady Eidson.

Touch events on iOS used to work with site isolated iframes, but with 290019@main
they stopped working because we are sending the wrong page identifier to a process,
which causes the lookup in EventDispatcher::dispatchTouchEvents to fail and it is
as if touch events never happened.  This fixes that by sending the correct PageIdentifier
for the process to which we are sending, and I also make EventDispatcher::dispatchTouchEvents
more robust by calling the completion handlers even if the page can't be found.
Also, capture a smarter pointer in EventDispatcher::touchEvent.

This is covered by the test http/tests/site-isolation/touch-events/touch-event-coordinates.html
which is only run on our internal bots, but it definitely started timing out after 290019@main.

* Source/WebKit/UIProcess/WebPageProxy.cpp:
(WebKit::WebPageProxy::sendPreventableTouchEvent):
(WebKit::WebPageProxy::sendUnpreventableTouchEvent):
* Source/WebKit/WebProcess/WebPage/EventDispatcher.cpp:
(WebKit::EventDispatcher::touchEvent):
(WebKit::EventDispatcher::dispatchTouchEvents):

Canonical link: https://commits.webkit.org/293870@main