REGRESSION(287602@main): [Debug] Networking process consistently crashes during Apple Pay sheet presentment on ENABLE_APPLE_PAY_REMOTE_UI builds https://bugs.webkit.org/show_bug.cgi?id=291739 rdar://149540206
Reviewed by Alex Christensen.

WebPaymentCoordinatorProxy lives in the networking process or in the UI
process, depending on whether ENABLE_APPLE_PAY_REMOTE_UI is true or not,
respectively. As such, it is incorrect to unconditionally claim that the
WebPaymentCoordinator messages are dispatched from the UI process.

This patch removes that claim and avoids debug assertions on _any_ Apple
Pay functionality on ENABLE_APPLE_PAY_REMOTE_UI platforms (i.e. iOS and
visionOS).

* Source/WebKit/WebProcess/ApplePay/WebPaymentCoordinator.messages.in:

Canonical link: https://commits.webkit.org/293876@main